+++ app/views/relationships/followers.html.erb
@@ -3,7 +3,7 @@
     <h2>Follower Users</h2>
     <% if @user == current_user %>
       <div>
-        <%= form_with url: user_followings_search_path(@user),method: :get, local: true do |f| %>
+        <%= form_with url: user_followers_search_path,method: :get, local: true do |f| %>
           <%= f.text_field :content %>
           <%= f.select :model,options_for_select({"User"=>"user","Give"=>"give","Take"=>"take"})%>
           <%= f.submit'検索' %>
@@ -14,6 +14,6 @@
       <%= render 'users/user', users: @users %>
     <% else %>
       <p>ユーザーはいません</p>
-      <% end %>
+    <% end %>
   </div>
 </div>
