--- app/views/searches/_search.html.erb
@@ -1,7 +1,0 @@
-<div class="serch_form">
-  <%= form_with url: searchs_path,method: :get, local: true do |f| %>
-    <%= f.text_field :content %>
-    <%= f.select :model,options_for_select({"User"=>"user","Give"=>"give","Take"=>"take"})%>
-    <%= f.submit'検索' %>
-  <% end %>
-</div>
