--- app/views/searches/search.html.erb
@@ -1,6 +1,0 @@
-<div class="container">
-  <div class="row">
-    <h2>Users</h2>
-    <%= render 'users/user', users: @records %>
-  </div>
-</div>
