+++ app/views/users/index.html.erb
@@ -2,7 +2,7 @@
   <div class="row">
     <h2>ユーザー一覧</h2>
     <div>
-      <%= form_with url: search_path,method: :get, local: true do |f| %>
+      <%= form_with url: user_search_path,method: :get, local: true do |f| %>
         <%= f.text_field :content %>
         <%= f.select :model,options_for_select({"User"=>"user","Give"=>"give","Take"=>"take"})%>
         <%= f.submit'検索' %>
